Update Antenna_Tuner.ino
Fixed the counter storage into the EEPROM so that it works. I migrated it to use byte*s so to chop up the longs to fit into and read out of the EEPROM. Not exactly the easiest solution to understand but I felt it was the quickest/easiest and most elegant solution for the problem.